% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{bind.socket}
\alias{bind.socket}
\title{Binds a socket to an address.}
\usage{
bind.socket(socket, address)
}
\arguments{
\item{socket}{Socket.}

\item{address}{Transport address to bind the socket to.}
}
\value{
`TRUE` if the socket was bound successfully, `FALSE` otherwise.
}
\description{
Binds a socket to an address.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
bind.socket(socket, "tcp://*:8080")

}
