% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{connect.socket}
\alias{connect.socket}
\title{Connects a socket to an address.}
\usage{
connect.socket(socket, address)
}
\arguments{
\item{socket}{Socket.}

\item{address}{Transport address to connect the socket to.}
}
\value{
`TRUE` if the socket connected successfully, `FALSE` otherwise.
}
\description{
Connects a socket to an address.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
connect.socket(socket, "tcp://*:8080")
}
