% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{disconnect.socket}
\alias{disconnect.socket}
\title{Disconnects a socket from an address.}
\usage{
disconnect.socket(socket, address)
}
\arguments{
\item{socket}{Socket.}

\item{address}{Transport address to disconnect the socket from}
}
\value{
`TRUE` if the socket connected successfully, `FALSE` otherwise.
}
\description{
Disconnects a socket from an address.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
connect.socket(socket, "tcp://*:8080")
disconnect.socket(socket, "tcp://*:8080")
}
