% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{get.last.endpoint}
\alias{get.last.endpoint}
\title{Retrieve the last endpoint bound for TCP and IPC transports
See ZMQ socket option `ZMQ_LAST_ENDPOINT`.}
\usage{
get.last.endpoint(socket)
}
\arguments{
\item{socket}{Socket.}
}
\value{
Last endpoint address.
}
\description{
Retrieve the last endpoint bound for TCP and IPC transports
See ZMQ socket option `ZMQ_LAST_ENDPOINT`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
connect.socket(socket, "tcp://*:8080")
get.last.endpoint(socket)
}
