% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{get.rcv.timeout}
\alias{get.rcv.timeout}
\title{Retrieves the receive timeout value.
See ZMQ socket option `ZMQ_RCVTIMEO`.}
\usage{
get.rcv.timeout(socket)
}
\arguments{
\item{socket}{Socket.}
}
\value{
Receive timeout in milliseconds.
-1 indicates an infinite timeout.
}
\description{
Retrieves the receive timeout value.
See ZMQ socket option `ZMQ_RCVTIMEO`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
get.rcv.timeout(socket)
}
