% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{get.rcvmore}
\alias{get.rcvmore}
\title{Gets whether the receive more flag is set on the socket.
See ZMQ socket option `ZMQ_RCVMORE`.}
\usage{
get.rcvmore(socket)
}
\arguments{
\item{socket}{Socket.}
}
\value{
Value of the receive more flag.
}
\description{
Gets whether the receive more flag is set on the socket.
See ZMQ socket option `ZMQ_RCVMORE`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
waiting <- get.rcvmore(socket)
}
