% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{get.send.timeout}
\alias{get.send.timeout}
\title{Retrieves the send timeout value.
See ZMQ socket option `ZMQ_SNDTIMEO`.}
\usage{
get.send.timeout(socket)
}
\arguments{
\item{socket}{Socket.}
}
\value{
Send timeout in milliseconds.
-1 indicates an infinite timeout.
}
\description{
Retrieves the send timeout value.
See ZMQ socket option `ZMQ_SNDTIMEO`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
get.send.timeout(socket)
}
