% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{init.context}
\alias{init.context}
\title{Initialize a ZMQ context.}
\usage{
init.context(threads = 1L)
}
\arguments{
\item{threads}{Number of threads for the context to use.}
}
\value{
ZMQ context.
}
\description{
Initialize a ZMQ context.
}
\examples{
init.context()
}
