% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{init.socket}
\alias{init.socket}
\title{Initializes a socket of the given type.}
\usage{
init.socket(context, socket.type)
}
\arguments{
\item{context}{ZMQ context.}

\item{socket.type}{Type of socket to initialize.}
}
\value{
ZMQ socket.
}
\description{
Initializes a socket of the given type.
}
\examples{
context <- init.context()
init.socket(context, "ZMQ_REQ")
}
