% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{receive.double}
\alias{receive.double}
\title{Receive a message as an double.}
\usage{
receive.double(socket)
}
\arguments{
\item{socket}{Socket.}
}
\value{
Received message as an double.
}
\description{
Receive a message as an double.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
connect.socket(socket, "tcp://*:8080")
msg <- receive.double(socket)
}
