% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{receive.int}
\alias{receive.int}
\title{Receive a message as an integer.}
\usage{
receive.int(socket)
}
\arguments{
\item{socket}{Socket.}
}
\value{
Received message as an integer.
}
\description{
Receive a message as an integer.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
connect.socket(socket, "tcp://*:8080")
msg <- receive.int(socket)
}
