% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{receive.multipart}
\alias{receive.multipart}
\title{Receive a multipart message.}
\usage{
receive.multipart(socket)
}
\arguments{
\item{socket}{}
}
\value{
List of message parts. Parts are unserialized.
}
\description{
Receive a multipart message.
}
\examples{
#' context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
connect.socket(socket, "tcp://*:8080")
send.socket(socket, "hello")
parts <- receive.multipart(socket)
}
