% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{receive.null.msg}
\alias{receive.null.msg}
\title{Receive a null message.}
\usage{
receive.null.msg(socket)
}
\arguments{
\item{socket}{Socket.}
}
\value{
Message.
}
\description{
Receive a null message.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
connect.socket(socket, "tcp://*:8080")
send.null.msg(socket)
receive.null.msg(socket)
}
