% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{receive.string}
\alias{receive.string}
\title{Receive a message as a string.}
\usage{
receive.string(socket)
}
\arguments{
\item{socket}{Socket.}
}
\value{
Received message as a string.
}
\description{
Receive a message as a string.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
connect.socket(socket, "tcp://*:8080")
msg <- receive.string(socket)
}
