% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{send.message.object}
\alias{send.message.object}
\title{Send a message.}
\usage{
send.message.object(socket, msg, send.more = FALSE)
}
\arguments{
\item{socket}{Socket.}

\item{msg}{Message.}

\item{send.more}{Whether more frames will be sent.}
}
\value{
`TRUE` if the message was sent successfully, `FALSE` otherwise.
}
\description{
Send a message.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
connect.socket(socket, "tcp://*:8080")
msg <- init.message("hello")
send.message.object(socket, msg)
send.message.object(socket, msg)
}
