% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{send.multipart}
\alias{send.multipart}
\title{Send a multipart message.
Uses `send.socket` with `serialize` set to `FALSE`.}
\usage{
send.multipart(socket, parts)
}
\arguments{
\item{socket}{Socket.}

\item{parts}{List of message parts.}
}
\value{
`TRUE` if the messages were sent successfully, `FALSE` otherwise.
}
\description{
Send a multipart message.
Uses `send.socket` with `serialize` set to `FALSE`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
connect.socket(socket, "tcp://*:8080")
msgs <- c("hello", "world")
send.multipart(socket, msgs)
}
