% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{send.null.msg}
\alias{send.null.msg}
\title{Sends a null message.}
\usage{
send.null.msg(socket, send.more = FALSE)
}
\arguments{
\item{socket}{Socket.}

\item{send.more}{Whether more frames will be sent.}
}
\value{
`TRUE` if the message was sent successfully, `FALSE` otherwise.
}
\description{
Sends a null message.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
connect.socket(socket, "tcp://*:8080")
send.null.msg(socket)
}
