% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{send.raw.string}
\alias{send.raw.string}
\title{Sends a raw string.}
\usage{
send.raw.string(socket, data, send.more = FALSE)
}
\arguments{
\item{socket}{Socket.}

\item{data}{String to send.}

\item{send.more}{Whether more frames in the message will be sent.}
}
\value{
`TRUE` if the message was sent successfully, `FALSE` otherwise.
}
\description{
Sends a raw string.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
connect.socket(socket, "tcp://*:8080")
send.raw.string(socket, "hello")
}
