% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{send.socket}
\alias{send.socket}
\title{Sends a message.}
\usage{
send.socket(
  socket,
  data,
  send.more = FALSE,
  serialize = TRUE,
  xdr = .Platform$endian == "big"
)
}
\arguments{
\item{socket}{Socket.}

\item{data}{Data to transmit.}

\item{send.more}{Whether more frames will be sent.}

\item{serialize}{Whether to call `serialize` before sending.}

\item{xdr}{Passed directly to `serialize` function.}
}
\value{
`TRUE` if the message was sent successfully, `FALSE` otherwise.
}
\description{
Sends a message.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
connect.socket(socket, "tcp://*:8080")
send.socket(socket, "hello")
}
