% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{set.affinity}
\alias{set.affinity}
\title{Set the I/O thread affinity for newly created connections of the socket.
See ZMQ socket option `ZMQ_AFFINITY`.}
\usage{
set.affinity(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Affinity.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set the I/O thread affinity for newly created connections of the socket.
See ZMQ socket option `ZMQ_AFFINITY`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
set.affinity(socket, 1L)
}
