% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{set.identity}
\alias{set.identity}
\title{Set the identity of the socket.
See ZMQ socket option `ZMQ_IDENTITY`.}
\usage{
set.identity(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Identity.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set the identity of the socket.
See ZMQ socket option `ZMQ_IDENTITY`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
set.identity(socket, "my_socket")
}
