% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{set.linger}
\alias{set.linger}
\title{Set the linger period.
See ZMQ socket option `ZMQ_LINGER`.}
\usage{
set.linger(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Linger period in milliseconds.
-1 specifies in infinite period.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set the linger period.
See ZMQ socket option `ZMQ_LINGER`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
set.linger(socket, 1L)
}
