% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{set.rate}
\alias{set.rate}
\title{Set the socket's multicast rate.
See ZMQ socket option `ZMQ_RATE`.}
\usage{
set.rate(socket, option.value)
}
\arguments{
\item{socket}{Socket. Must be multicast.}

\item{option.value}{Rate in kilobytes per second.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set the socket's multicast rate.
See ZMQ socket option `ZMQ_RATE`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_PUB")
set.rate(socket, 100L)
}
