% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{set.rcv.hwm}
\alias{set.rcv.hwm}
\title{Set high water mark for inbound messages.
See ZMQ socket option `ZMQ_RCVHWM`.}
\usage{
set.rcv.hwm(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Maximum queued messages.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set high water mark for inbound messages.
See ZMQ socket option `ZMQ_RCVHWM`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
set.rcv.hwm(socket, 100L)
}
