% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{set.rcv.timeout}
\alias{set.rcv.timeout}
\title{Sets the timeout for receive operations.
See ZMQ socket option `ZMQ_RCVTIMEO`.}
\usage{
set.rcv.timeout(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Timeout in milliseconds.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Sets the timeout for receive operations.
See ZMQ socket option `ZMQ_RCVTIMEO`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
set.rcv.timeout(socket, 1000L)
}
