% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{set.rcvbuf}
\alias{set.rcvbuf}
\title{Set the kernel receive buffer size.
See ZMQ socket option `ZMQ_RCVBUF`.}
\usage{
set.rcvbuf(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Buffer size in bytes.
-1 specifies the OS default should be used.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set the kernel receive buffer size.
See ZMQ socket option `ZMQ_RCVBUF`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
set.rcvbuf(socket, -1L)
}
