% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{set.reconnect.ivl.max}
\alias{set.reconnect.ivl.max}
\title{Set maximum reconnection interval.
See ZMQ socket option `ZMQ_RECONNECT_IVL_MAX`.}
\usage{
set.reconnect.ivl.max(socket, option.value)
}
\arguments{
\item{socket}{Socket. Must be connection-oriented.}

\item{option.value}{Maximum reconnect interval in milliseconds.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set maximum reconnection interval.
See ZMQ socket option `ZMQ_RECONNECT_IVL_MAX`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
set.reconnect.ivl.max(socket, 1000L)
}
