% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{set.reconnect.ivl}
\alias{set.reconnect.ivl}
\title{Set the reconnection interval.
See ZMQ socket option `ZMQ_RECONNECT_IVL`.}
\usage{
set.reconnect.ivl(socket, option.value)
}
\arguments{
\item{socket}{Sokcet.}

\item{option.value}{Reconnection interval in milliseconds.
-1 specifies no reconnection.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set the reconnection interval.
See ZMQ socket option `ZMQ_RECONNECT_IVL`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
set.reconnect.ivl(socket, 100L)
}
