% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{set.send.hwm}
\alias{set.send.hwm}
\title{Set high water mark for outbound messages.
See ZMQ socket option `ZMQ_SNDHWM`.}
\usage{
set.send.hwm(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Maximum queued messages.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set high water mark for outbound messages.
See ZMQ socket option `ZMQ_SNDHWM`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
set.send.hwm(socket, 100L)
}
