% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{set.send.timeout}
\alias{set.send.timeout}
\title{Sets the timeout for send operations.
See ZMQ socket option `ZMQ_SNDTIMEO`.}
\usage{
set.send.timeout(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Timeout in milliseconds.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Sets the timeout for send operations.
See ZMQ socket option `ZMQ_SNDTIMEO`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
set.send.timeout(socket, 1000L)
}
