% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{set.sndbuf}
\alias{set.sndbuf}
\title{Set the kernel transmit buffer size.
See ZMQ socket option `ZMQ_SNDBUF`.}
\usage{
set.sndbuf(socket, option.value)
}
\arguments{
\item{socket}{Socket.}

\item{option.value}{Buffer size in bytes.
-1 specifies the OS default should be used.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set the kernel transmit buffer size.
See ZMQ socket option `ZMQ_SNDBUF`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
set.sndbuf(socket, -1L)
}
