% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{set.tcp.keepalive}
\alias{set.tcp.keepalive}
\title{Sets the TCP keep alive option.}
\usage{
set.tcp.keepalive(
  socket,
  active = -1L,
  idle = -1L,
  count = -1L,
  interval = -1L
)
}
\arguments{
\item{socket}{Socket.}

\item{active}{Whether to activate (1L), inactivate (0L) or use OS default (-1L).
See ZMQ socket option `ZMQ_TCP_KEEPALIVE`.}

\item{idle}{Idle time until keep-alive is sent (use OS default: -1L).
See ZMQ socket option `ZMQ_TCP_KEEPALIVE_IDLE`.}

\item{count}{Number of probes missed to drop connection (use OS default: -1L).
See ZMQ socket option `ZMQ_TCP_KEEPALIVE_CNT`.}

\item{interval}{Time between keepalive probes (use OS default: -1L).
See ZMQ socket option `ZMQ_TCP_KEEPALIVE_INTVL`}
}
\value{
`TRUE` if the options were set successfully, `FALSE` otherwise.
}
\description{
Sets the TCP keep alive option.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
set.tcp.keepalive(socket)
}
