% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{set.zmq.backlog}
\alias{set.zmq.backlog}
\title{Set maximum length of the queue of outstanding connections
See ZMQ socket option `ZMQ_BACKLOG`.}
\usage{
set.zmq.backlog(socket, option.value)
}
\arguments{
\item{socket}{Socket. Must be connection-oriented.}

\item{option.value}{Number of connections.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Set maximum length of the queue of outstanding connections
See ZMQ socket option `ZMQ_BACKLOG`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REP")
set.zmq.backlog(socket, 100L)
}
