% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{subscribe}
\alias{subscribe}
\title{Establish a new message filter on a `ZMQ_SUB` socket.
See ZMQ socket option `ZMQ_SUBSCRIBE`.}
\usage{
subscribe(socket, option.value)
}
\arguments{
\item{socket}{Socket. Must be of type `ZMQ_SUB`.}

\item{option.value}{Filter.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Establish a new message filter on a `ZMQ_SUB` socket.
See ZMQ socket option `ZMQ_SUBSCRIBE`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_SUB")
subscribe(socket)
}
