% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{unsubscribe}
\alias{unsubscribe}
\title{Unsubscribe from a message filter.
See ZMQ socket option `ZMQ_UNSUBSCRIBE`.}
\usage{
unsubscribe(socket, option.value)
}
\arguments{
\item{socket}{Socket. Must be of type `ZMQ_SUB`.}

\item{option.value}{Filter.}
}
\value{
`TRUE` if the option was set successfully, `FALSE` otherwise.
}
\description{
Unsubscribe from a message filter.
See ZMQ socket option `ZMQ_UNSUBSCRIBE`.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_SUB")
subscribe(socket)
unsubscribe(socket, "ignore")
}
