% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{zmq.errno}
\alias{zmq.errno}
\title{Returns the latest error code from ZMQ.}
\usage{
zmq.errno()
}
\value{
Error code.
}
\description{
Returns the latest error code from ZMQ.
}
\examples{
zmq.errno()
}
