% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{zmq.strerror}
\alias{zmq.strerror}
\title{Returns the latest error message from ZMQ.}
\usage{
zmq.strerror()
}
\value{
Error message.
}
\description{
Returns the latest error message from ZMQ.
}
\examples{
zmq.strerror()
}
