% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq.R
\name{init.message}
\alias{init.message}
\title{Creates a ZMQ message that can be sent multiple times.}
\usage{
init.message(data, serialize = TRUE, xdr = .Platform$endian == "big")
}
\arguments{
\item{data}{Data to transmit.}

\item{serialize}{Whether to call `serialize` before sending.}

\item{xdr}{Passed directly to `serialize` function.}
}
\value{
ZMQ message.
}
\description{
Creates a ZMQ message that can be sent multiple times.
}
\examples{
init.message("hello")
}
