% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzmq-package.R
\name{receive.socket}
\alias{receive.socket}
\title{Receive a message.}
\usage{
receive.socket(socket, unserialize = TRUE, dont.wait = FALSE)
}
\arguments{
\item{socket}{Socket.}

\item{unserialize}{Whether to unserialize the message.}

\item{dont.wait}{`FALSE` to block, `TRUE` for non-blocking.}
}
\value{
Message.
}
\description{
Receive a message.
}
\examples{
context <- init.context()
socket <- init.socket(context, "ZMQ_REQ")
connect.socket(socket, "tcp://*:8080")
send.socket(socket, "hello")
msg <- receive.socket(socket)
}
